Flatten folder slightly
I'm not sure how this project should be structured yet, but having
unnecessary depth is not the thing to aim for.
