+++ packages/harpstrata/src/harpstrata/IsActive/ActiveIds/index.ts
@@ -1,4 +1,1 @@
 export { getActiveIdsPair } from './getActiveIdsPair'
-export type { ActiveDegreeIds, ActivePitchIds, ActiveIds } from './types'
-
-export { EXAMPLE_IS_ACTIVE_IDS_PAIR } from './testResources'
