+++ packages/harpstrata/src/harpstrata/IsActive/index.ts
@@ -1,7 +1,6 @@
 export { getActiveIdsPair } from './ActiveIds'
 
-export type { ActiveDegreeIds, ActivePitchIds, ActiveIds } from './ActiveIds'
-
+export type { ActiveDegreeIds, ActivePitchIds, ActiveIds } from './types'
 export { IsActiveIds } from './types'
 
-export { EXAMPLE_IS_ACTIVE_IDS_PAIR } from './ActiveIds'
+export { EXAMPLE_IS_ACTIVE_IDS_PAIR } from './testResources'
