+++ packages/harpstrata/src/harpstrata/IsActive/testResources/index.ts
@@ -1,6 +1,6 @@
 import type { ActivePitchIds, ActiveDegreeIds } from '../types'
-import { PitchIds } from '../../../Pitch'
-import { DegreeIds } from '../../../Degree'
+import { PitchIds } from '../../Pitch'
+import { DegreeIds } from '../../Degree'
 
 const MAJOR_PENTATONIC_SCALE_DEGREE_IDS: ActiveDegreeIds = [
   DegreeIds.Root,
