+++ packages/harpstrata/src/harpstrata/IsActive/types/index.ts
@@ -1,6 +1,17 @@
-import type { ActiveIds } from '../ActiveIds'
 import type { PitchMatrix } from '../../Pitch'
+import { PitchIds } from '../../Pitch'
+import { DegreeIds } from '../../Degree'
 import type { DegreeMatrix } from '../../Degree'
+
+export type ActiveDegreeIds = ReadonlyArray<DegreeIds>
+export type ActivePitchIds = ReadonlyArray<PitchIds>
+
+export type ActiveIds = ActivePitchIds | ActiveDegreeIds
+
+export type ActiveIdsPair = {
+  readonly activeDegreeIds: ActiveDegreeIds
+  readonly activePitchIds: ActivePitchIds
+}
 
 export type IsActiveProps = {
   degreeMatrix: DegreeMatrix
